Create my_each_with_index, my_select and my_all?
